# paise-wallet/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Paise Wallet API
  description: |
    Wallet / P2P transfer service. Money is held as integer paise. All business
    endpoints require a Bearer JWT; the caller's identity is always the verified
    JWT subject, never a header or body field.
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /dev/token:
    get:
      tags: [dev]
      operationId: devToken
      summary: Mint a development JWT for a user (dev only)
      security: []
      parameters:
        - name: user
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A one-hour HS256 JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/Error'
  /dev/fund:
    post:
      tags: [dev]
      operationId: devFund
      summary: Credit a user's wallet for testing (dev only)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundRequest'
      responses:
        '200':
          description: The funded wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    AccountResponse:
      type: object
      required: [user_id, wallet_id, balance_paise]
      properties:
        user_id:
          type: string
        wallet_id:
          type: string
          format: uuid
        balance_paise:
          type: integer
          format: int64
...
    ErrorResponse:
      type: object
      required: [error_code, message, request_id]
      properties:
        error_code:
          type: string
        message:
          type: string
        request_id:
          type: string
    TokenResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    FundRequest:
      type: object
      required: [user_id, amount_paise]
      properties:
        user_id:
          type: string
          minLength: 1
        amount_paise:
          type: integer
          format: int64
          minimum: 1